requirements.txt
boto3==1.28.8       # AWS SDK for Python, necesario para interactuar con AWS
paramiko==2.11.0    # Biblioteca para manejar SSH, útil para instalación remota en Linux
pywinrm==0.4.2      # Biblioteca para manejar conexiones remotas a Windows a través de WinRM
